Move write method to drawer pkg
