+++ drawer/drawer_test.go
@@ -4,6 +4,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/gdamore/tcell"
 	"github.com/wasanx25/goss/drawer"
 	"github.com/wasanx25/goss/event"
 )
@@ -91,7 +92,7 @@ func TestAddPosition(t *testing.T) {
 	}
 
 	d := drawer.New("", 0, 0, 1) // dummy args
-	d.Reset()
+	d.PositionReset()
 	for _, tt := range tests {
 		d.AddPosition(tt.args)
 
@@ -105,13 +106,13 @@ func TestAddPosition(t *testing.T) {
 			t.Errorf(string(tt.args)+" is expected=%d, got=%d", tt.expectRow, row)
 		}
 
-		d.Reset() // reset
+		d.PositionReset() // reset
 	}
 }
 
 func TestInitPosition(t *testing.T) {
 	d := drawer.New("", 0, 0, 1) // dummy args
-	d.Reset()
+	d.PositionReset()
 
 	col, row := d.Position()
 
@@ -137,4 +138,51 @@ func TestBreak(t *testing.T) {
 	if row != 1 {
 		t.Errorf("expected=%d, got=%d", 1, row)
 	}
+}
+
+func TestWrite(t *testing.T) {
+	tui := tcell.NewSimulationScreen("")
+	if err := tui.Init(); err != nil {
+		t.Fatal(err)
+	}
+
+	d := drawer.New("test1\ntest2 test3\ttest4\ntest5", 0, 2, 5)
+	d.SetLimit(5)
+
+	tui.SetSize(30, 5)
+
+	d.Write(tui, tcell.StyleDefault, tcell.StyleDefault)
+
+	tui.Show()
+
+	// trim end space if use heredoc
+	slice := []string{
+		" 1   test1                    \n",
+		" 2   test2  test3    test4    \n",
+		" 3   test5                    \n",
+		"                              \n",
+		"                              \n",
+	}
+	expected := slice[0] + slice[1] + slice[2] + slice[3] + slice[4]
+
+	actual := getString(tui)
+
+	if actual != expected {
+		t.Errorf("expected=%v, got=%v", expected, actual)
+	}
+}
+
+func getString(tui tcell.SimulationScreen) string {
+	width, _ := tui.Size()
+	cells, _, _ := tui.GetContents()
+
+	var runes []rune
+	for i, c := range cells {
+		runes = append(runes, c.Runes...)
+		if (i+1)%width == 0 {
+			runes = append(runes, '\n')
+		}
+	}
+
+	return string(runes)
 }
