--- viewer/export_test.go
@@ -1,4 +1,0 @@
-package viewer
-
-var Write = (*Viewer).write
-var SetLimit = (*Viewer).setLimit
